refactor(backend): colocate appeal domain into feature package
Moves the appeal controller, entity, repository and DTOs out of the
generic controller/database/dto layer packages into
net.onelitefeather.blackhole.backend.appeal, alongside the
already-colocated AppealDecisionService/AppealEligibilityService.
Pure physical move, no behavior change.
