+++ backend/src/main/java/net/onelitefeather/blackhole/backend/appeal/AppealEligibilityService.java
@@ -2,11 +2,8 @@ package net.onelitefeather.blackhole.backend.appeal;
 
 import io.micronaut.context.annotation.Value;
 import jakarta.inject.Singleton;
-import net.onelitefeather.blackhole.backend.database.entities.AppealEntity;
 import net.onelitefeather.blackhole.backend.database.entities.EloProfileEntity;
-import net.onelitefeather.blackhole.backend.database.repository.AppealRepository;
 import net.onelitefeather.blackhole.backend.database.repository.EloProfileRepository;
-import net.onelitefeather.blackhole.backend.dto.AppealStatus;
 import net.onelitefeather.blackhole.backend.dto.EloTrack;
 import net.onelitefeather.blackhole.backend.elo.EloService;
 import net.onelitefeather.blackhole.backend.punishment.PunishmentEntity;
