+++ backend/src/main/java/net/onelitefeather/blackhole/backend/appeal/AppealEntity.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.database.entities;
+package net.onelitefeather.blackhole.backend.appeal;
 
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.persistence.Column;
@@ -11,8 +11,6 @@ import jakarta.persistence.Index;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import net.onelitefeather.blackhole.backend.database.converter.MapStringObjectConverter;
-import net.onelitefeather.blackhole.backend.dto.AppealDTO;
-import net.onelitefeather.blackhole.backend.dto.AppealStatus;
 import net.onelitefeather.blackhole.backend.punishment.PunishmentEntity;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
