+++ backend/src/main/java/net/onelitefeather/blackhole/backend/appeal/AppealSubmissionDTO.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.dto;
+package net.onelitefeather.blackhole.backend.appeal;
 
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.ReflectiveAccess;
